Added a Driver to the Speedway API.

+++ src/main/java/com/galvanize/speedway/entities/Racecar.java
@@ -1,13 +1,12 @@
 package com.galvanize.speedway.entities;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity
 @Data
@@ -26,8 +25,11 @@ public class Racecar {
 
     private int year;
 
-    private int owner;
+
+    private Long owner;
 
     private int topSpeed;
+
+
 
 }
